pyproject.toml
[project]
name = "pathfinder"
version = "0.1.0"
description = "Scenic driving route finder using percolation theory and semantic segmentation"
requires-python = ">=3.9"
dependencies = [
    "seaborn",
    "matplotlib",
    "numpy",
    "geopandas",
    "shapely",
    "tqdm",
    "requests",
    "fpdf2",
    "simplekml",
    "folium",
]

[project.optional-dependencies]
scraping = ["selenium"]
dev = [
    "pytest",
    "pytest-cov",
]

# ML inference deps (torch, mmcv-full, mmsegmentation, opencv-python)
# are environment-specific (CUDA, etc.) and must be installed manually.
# See scenery/README.md for instructions.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "scenery", "data_pipeline"]
exclude = ["Data", "data", "output", "archive", "papers", "Navigate", "Pathfinder", "Explore", "*.png", "*.h5", "*.pth"]

[tool.pytest.ini_options]
testpaths = ["tests"]
